refactor: extract duplicated PNG test helpers to shared utility (#546)
